+++ HISTORY.rst
@@ -13,6 +13,7 @@ History
 * Added missing requirement 'gdal'.
 * Bumped version of CI base image.
 * Updated version pinnings of enmapbox requirements.
+* Improved test output in case POLYMER is not properly installed.
 
 
 0.19.4 (2023-08-09)